+++ langs/csharp/tests/VMx.Tests/Commands/ConfirmationDecoratorCommandTests.cs
@@ -1,4 +1,6 @@
 using System.Reactive.Linq;
+using System.Reactive.Subjects;
+using System.Reflection;
 using FluentAssertions;
 using VMx.Commands;
 using Xunit;
@@ -7,6 +9,27 @@ namespace VMx.Tests.Commands;
 
 public class ConfirmationDecoratorCommandTests
 {
+    [Fact]
+    public void Dispose_Disposes_Error_Channel_When_Completion_Observer_Throws()
+    {
+        var inner = RelayCommand.Builder().Build();
+        var command = new ConfirmationDecoratorCommand(
+            inner,
+            () => Task.FromResult(true));
+        using var subscription = command.Errors.Subscribe(
+            _ => { },
+            () => throw new InvalidOperationException("terminal observer"));
+
+        Action dispose = command.Dispose;
+
+        dispose.Should().Throw<InvalidOperationException>().WithMessage("terminal observer");
+        var errors = (Subject<Exception>)typeof(ConfirmationDecoratorCommand)
+            .GetField("_errors", BindingFlags.Instance | BindingFlags.NonPublic)!
+            .GetValue(command)!;
+        errors.IsDisposed.Should().BeTrue();
+        dispose.Should().NotThrow("all terminal cleanup completed before rethrowing");
+    }
+
     [Fact]
     public async Task Dispose_Waits_For_InFlight_Error_Delivery()
     {