+++ langs/csharp/tests/VMx.Tests/Components/ComponentVMTests.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using System.Reflection;
 using FluentAssertions;
+using VMx.Commands;
 using VMx.Components;
 using VMx.Lifecycle;
 using VMx.Messages;
@@ -42,6 +43,31 @@ public class ComponentVMTests
             .GetField("_triggerDisposed", BindingFlags.Instance | BindingFlags.NonPublic)!
             .GetValue(vm);
         triggerDisposed.Should().Be(true);
+        vm.Dispose();
+    }
+
+    [Fact]
+    public void Throwing_Terminal_Command_Observer_Still_Completes_Base_Teardown()
+    {
+        var (vm, _, _) = BuildVm();
+        var select = (RelayCommand)vm.SelectCommand;
+        var deselect = (RelayCommand)vm.DeselectCommand;
+        select.CanExecuteChanged += (_, _) =>
+            throw new InvalidOperationException("terminal observer failure");
+
+        Action dispose = vm.Dispose;
+
+        dispose.Should().Throw<InvalidOperationException>()
+            .WithMessage("terminal observer failure");
+        vm.Status.Should().Be(ConstructionStatus.Disposed);
+        var triggerDisposed = typeof(ComponentVMBase)
+            .GetField("_triggerDisposed", BindingFlags.Instance | BindingFlags.NonPublic)!
+            .GetValue(vm);
+        triggerDisposed.Should().Be(true);
+        typeof(RelayCommand).GetField("_disposed", BindingFlags.Instance | BindingFlags.NonPublic)!
+            .GetValue(select).Should().Be(true);
+        typeof(RelayCommand).GetField("_disposed", BindingFlags.Instance | BindingFlags.NonPublic)!
+            .GetValue(deselect).Should().Be(true);
         vm.Dispose();
     }